requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
whitenoise==6.6.0
python-dotenv==1.0.0
firebase-admin==6.2.0
